retry releases (#4913)
This PR changes the release version format for snapshot releases to
allow for an optional "build number", i.e. "how many times have we
screwed up this release before this attempt?".

Adding this to the version string should be fine because:

- It is a number, so GHC will be happy.
- It is dot-separated and (manually) incrementing, so all three formats
  will sort it correctly.
- The SemVer->GHC conversion only looks at the beginning and ending of
  the snapshot string, and this changes the middle.

This is necessary because sometimes we screw up releases (e.g. #4902),
but also because sometimes releases screw themselves up by somehow
corrupting the Windows cache, and so far changing the version number has
been the only way out of that. So far that has meant changing the target
commit, but that's a very poor reason to choose a target commit.

We should not have to include additional code in a release just because
our release process is flaky.

CHANGELOG_BEGIN
CHANGELOG_END