Introduce DamlSubmissionBatch and the BatchingLedgerWriter (#4964)
* Introduce DamlSubmissionBatch and the BatchingLedgerWriter

This introduces the DamlSubmissionBatch message to group
submissions into a single message and extends Envelope to
carry the batch. We're using the envelope wrapping for consistency
and compatibility.

We're adding this to kvutils version 3 as it has not yet been released
into the wild and as this is not backwards incompatible change.

Support for batching is implemented with the BatchingLedgerWriter
that wraps a LedgerWriter and groups submissions into a batch based
on size and time duration.

For implementing the validation of a batch we will require some rework
in the SubmissionValidator to be able to produce multiple "LogResult"s,
e.g. commit on the in-memory ledger results in an "Index" which is used
to signal new head to dispatcher. With a batch we'd need to pick max index.

CHANGELOG_BEGIN
CHANGELOG_END

* Add missing copyright header

* Address code reviews

* Post rebase fixes

* Rename BATCH -> SUBMISSION_BATCH

* Address code reviews, add further tests and cleanup.

* Add test for DefaultBatchingQueue.close

* Use generous timeouts

* Renamed BatchMessage => SubmissionBatchMessage. Added default boolean parameter value. Added simple test case.

* Removed unused include.

* Address final code review

Co-authored-by: Miklos Erdelyi <miklos.erdelyi@digitalasset.com>