damlc: Drop support for DAML-LF 1.5 and 1.6
The sandbox has already dropped support for everything before
DAML-LF 1.7, hence there's no point in being able to produce or consume
them with the compiler.

This PR only drops support for DARs in DAML-LF 1.5 or 1.6 and support
for accepting DAML-LF 1.6 as the `--target`.

The code paths through the compiler which were only used for these
DAML-LF versions will be dropped in a followup PR.

CHANGELOG_BEGIN
CHANGELOG_END