add comment to intro code
Having just spent a bit of time walking a user through this example on
Slack, I believe the use of a negative code sample here (`Transfer`
seems a bit useless) is confusing to newcomers. The user had written
code with the same structure as `Iou` and `Transfer` and did not
understand why their code did not work, as they had done it "just like
in the documentation".

Hopefully adding this warning helps, though maybe the structure of this
section should be rethought. That user also had trouble figuring out
where the `Mutual_Transfer` choice fits in with the rest, and I had to
come look at the raw code here to figure it out. Hopefully adding a
little bit of context here will also clarify.

CHANGELOG_BEGIN
CHANGELOG_END