TS libs: Fail linting on warnings (#5332)
Currently, we happily accept warnings of the linter as if nothing was
wrong. This PR makes the linter fail on warnings. It also moves the
linter configs into their own files to unclutter the `package.json`
files we ship to our users a bit.

Ane we fix all the warnings we now get from the linter.

CHANGELOG_BEGIN
CHANGELOG_END