daml-lf: Tiny performance improvements. (#5493)
* daml-lf/data: Improve performance of `areEqual`.

`areEqual` constructs a one-element `FrontStack`, which should be cheap,
but turns out to be expensive because it constructs builders along the
way. This provides overloads for `apply` that construct the FrontStack
directly, rather than jumping through hoops.

This also changes the behavior of `FrontStack#equals` to check the
length first, which makes the tests work but might also help
performance.

* daml-lf/data: Merge `length` and `len`.

In `FrontStack`, `BackStack`, and `ImmArray`.

No reason to have `def length = len` when `len` can just be public.

* daml-lf/interpreter: Allocate the right-sized list for arguments.

We spend a fair amount of time in `ArrayList#add`; this lessens that by
making sure than when we clone the `args` array for function
application, we leave some room for adding the arguments.

CHANGELOG_BEGIN
CHANGELOG_END

* daml-lf/data: Revert accidentally-lengthened imports.

Co-Authored-By: Stephen Compall <stephen.compall@daml.com>

* daml-lf/interpreter: Rename `functionArgs` to `extendedArgs`.

Co-authored-by: Stephen Compall <stephen.compall@daml.com>