create-daml-app test: Move back to daml start
Previously we changed to running the sandbox and JSON API separately to
have more control over port allocation. Now the same behaviour is
possible using `daml start`, which is preferable because it's what we
suggest users of the Getting Started Guide should use. This change
returns to using `daml start` in the end-to-end test. We use
`--sandbox-port=0` to get a dynamically allocated sandbox port which is
passed as the JSON API ledger port. We continue to hardcode the JSON API port.

changelog_begin
changelog_end